+++ src/lib/AsyncTypeaheadInput.tsx
@@ -12,7 +12,7 @@ import {
 import { useDebounceHook } from "./hooks/useDebounceHook";
 import { useSafeNameId } from "./hooks/useSafeNameId";
 import { useFormContext } from "./context/FormContext";
-import { TypeaheadTextField } from "./TypeaheadTextField";
+import { TypeaheadTextField } from "./components/Typeahead/TypeaheadTextField";
 import { FormGroupLayout } from "./FormGroupLayout";
 
 interface AsyncTypeaheadInputRef {
@@ -128,7 +128,7 @@ const AsyncTypeaheadInput = <T extends FieldValues>(props: AsyncTypeaheadInputPr
       name={name}
       label={useBootstrapStyle ? label : undefined}
       labelStyle={useBootstrapStyle ? { color: "#8493A5", fontSize: 14 } : undefined}
-      layout="typeahead"
+      layout="muiInput"
     >
       <Autocomplete<TypeaheadOption, boolean, boolean, boolean>
         {...autocompleteProps}